requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
